Moved Web3Context stuff from @metafam/web to @metafam/utils package to be shared by other packages. Fixes #295
