+++ packages/utils/package.json
@@ -13,9 +13,12 @@
     "precommit": "yarn lint-staged"
   },
   "dependencies": {
+    "@walletconnect/web3-provider": "^1.2.1",
     "ethers": "^5.0.8",
     "js-base64": "^3.5.2",
-    "uuid": "^8.0.0"
+    "react": "^16.13.1",
+    "uuid": "^8.0.0",
+    "web3modal": "^1.9.0"
   },
   "devDependencies": {
     "@types/uuid": "^8.0.0"
