+++ packages/utils/src/index.ts
@@ -3,3 +3,5 @@ import * as did from './did';
 export { did };
 export * from './arrayHelpers';
 export * from './promiseHelpers';
+export * from './context/Web3Context';
+export * from './lib/auth';
