+++ packages/utils/src/lib/auth.ts
